+++ app/src/main/java/woowacourse/movie/fragment/movielist/adapter/MovieRecyclerViewAdapter.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.fragment.movielist
+package woowacourse.movie.fragment.movielist.adapter
 
 import android.view.LayoutInflater
 import android.view.ViewGroup
