+++ app/src/main/java/woowacourse/movie/fragment/movielist/adapter/MovieRecyclerViewHolder.kt
@@ -1,4 +1,4 @@
-package woowacourse.movie.fragment.movielist
+package woowacourse.movie.fragment.movielist.adapter
 
 import android.view.View
 import android.widget.Button
